Add support for scoped deadlines and timeouts (#30)
This adds support for timeouts/deadlines in duet scopes. Within a scope with a deadline, any await on a future that does not complete by the deadline will raise a TimeoutError. Deadlines in a scope are also inherited by tasks spawned within that scope. In addition, nested scopes can have their own deadlines though the effective deadline will always be the min of the scope's own deadline and the parent scope's deadline.

Fixes #28
Review: @kunalq