Add support for contextvars in duet tasks
When creating new tasks (e.g. with scope.spawn) we copy the current
`contextvars.Context` and then use that context whenever we advance the
coroutine for that task. This ensures that changes made to context vars
in a given task will be visible to that task, and only that task,
whenever it is running.